+++ CHANGELOG.md
@@ -19,6 +19,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Change FileDriver to be optional
 - Enforce https on ISTAT source
 - Remove DataCursor from API
+- Refactor SdmxManager API
 
 ### Fixed